requirements.txt
Django==1.6.3
Pygments==2.0.1
South==1.0.1
bpython==0.13.1
djangorestframework==3.0.2
gnureadline==6.3.3
ipdb==0.8
ipython==2.3.1
wsgiref==0.1.2
